Year added to DateRegister
Year and weekday code are excluded in the comparison between last
stored time and the actual timestamp. If the date is not equal or UART
connection initialised StoreDateTime function is called.  In the
function date register (RTC_DR) is queried and only the year added to time
stamp date. DateRegister remained compact.